Move benchmarking code from Types.hs to Benchmark.hs
