+++ Gauge/Types.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE CPP         #-}
-{-# LANGUAGE RankNTypes  #-}
-{-# LANGUAGE DeriveDataTypeable, DeriveGeneric, GADTs, RecordWildCards #-}
+{-# LANGUAGE DeriveDataTypeable, DeriveGeneric, RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections       #-}
 {-# OPTIONS_GHC -funbox-strict-fields #-}
@@ -35,9 +34,6 @@ module Gauge.Types
     , DisplayMode(..)
     , MatchType(..)
     , Verbosity(..)
-    -- * Benchmark descriptions
-    , Benchmarkable(..)
-    , Benchmark(..)
     -- * Measurements
     , Measured(..)
     , fromInt
@@ -51,30 +47,12 @@ module Gauge.Types
     , measureKeys
     , rescale
     , secs
-    -- * Benchmark construction
-    , env
-    , envWithCleanup
-    , perBatchEnv
-    , perBatchEnvWithCleanup
-    , perRunEnv
-    , perRunEnvWithCleanup
-    , toBenchmarkable
-    , bench
-    , bgroup
-    , addPrefix
-    , benchNames
-    -- ** Evaluation control
-    , whnf
-    , nf
-    , nfIO
-    , whnfIO
     ) where
 
 -- Temporary: to support pre-AMP GHC 7.8.4:
 import Control.Applicative
 
 import Control.DeepSeq (NFData(rnf))
-import Control.Exception (evaluate)
 import Control.Monad (when, unless)
 import Data.Data (Data, Typeable)
 import Data.Int (Int64)
@@ -169,28 +147,6 @@ data Config = Config {
       -- ^ Mode of operation
     , displayMode  :: DisplayMode
     } deriving (Eq, Read, Show, Typeable, Data, Generic)
-
-
--- | A pure function or impure action that can be benchmarked. The
--- 'Int64' parameter indicates the number of times to run the given
--- function or action.
-data Benchmarkable = forall a . NFData a =>
-    Benchmarkable
-      { allocEnv :: Int64 -> IO a
-      , cleanEnv :: Int64 -> a -> IO ()
-      , runRepeatedly :: a -> Int64 -> IO ()
-      , perRun :: Bool
-      }
-
-noop :: Monad m => a -> m ()
-noop = const $ return ()
-{-# INLINE noop #-}
-
--- | Construct a 'Benchmarkable' value from an impure action, where the 'Int64'
--- parameter indicates the number of times to run the action.
-toBenchmarkable :: (Int64 -> IO ()) -> Benchmarkable
-toBenchmarkable f = Benchmarkable noop (const noop) (const f) False
-{-# INLINE toBenchmarkable #-}
 
 -- | A collection of measurements made while benchmarking.
 --
@@ -449,281 +405,3 @@ fromDouble d | isInfinite d || isNaN d = Nothing
 toDouble :: Maybe Double -> Double
 toDouble Nothing  = -1/0
 toDouble (Just d) = d
-
--- | Apply an argument to a function, and evaluate the result to weak
--- head normal form (WHNF).
-whnf :: (a -> b) -> a -> Benchmarkable
-whnf = pureFunc id
-{-# INLINE whnf #-}
-
--- | Apply an argument to a function, and evaluate the result to
--- normal form (NF).
-nf :: NFData b => (a -> b) -> a -> Benchmarkable
-nf = pureFunc rnf
-{-# INLINE nf #-}
-
-pureFunc :: (b -> c) -> (a -> b) -> a -> Benchmarkable
-pureFunc reduce f0 x0 = toBenchmarkable (go f0 x0)
-  where go f x n
-          | n <= 0    = return ()
-          | otherwise = evaluate (reduce (f x)) >> go f x (n-1)
-{-# INLINE pureFunc #-}
-
--- | Perform an action, then evaluate its result to normal form.
--- This is particularly useful for forcing a lazy 'IO' action to be
--- completely performed.
-nfIO :: NFData a => IO a -> Benchmarkable
-nfIO = toBenchmarkable . impure rnf
-{-# INLINE nfIO #-}
-
--- | Perform an action, then evaluate its result to weak head normal
--- form (WHNF).  This is useful for forcing an 'IO' action whose result
--- is an expression to be evaluated down to a more useful value.
-whnfIO :: IO a -> Benchmarkable
-whnfIO = toBenchmarkable . impure id
-{-# INLINE whnfIO #-}
-
-impure :: (a -> b) -> IO a -> Int64 -> IO ()
-impure strategy a = go
-  where go n
-          | n <= 0    = return ()
-          | otherwise = a >>= (evaluate . strategy) >> go (n-1)
-{-# INLINE impure #-}
-
--- | Specification of a collection of benchmarks and environments. A
--- benchmark may consist of:
---
--- * An environment that creates input data for benchmarks, created
---   with 'env'.
---
--- * A single 'Benchmarkable' item with a name, created with 'bench'.
---
--- * A (possibly nested) group of 'Benchmark's, created with 'bgroup'.
-data Benchmark where
-    Environment  :: NFData env
-                 => IO env -> (env -> IO a) -> (env -> Benchmark) -> Benchmark
-    Benchmark    :: String -> Benchmarkable -> Benchmark
-    BenchGroup   :: String -> [Benchmark] -> Benchmark
-
--- | Run a benchmark (or collection of benchmarks) in the given
--- environment.  The purpose of an environment is to lazily create
--- input data to pass to the functions that will be benchmarked.
---
--- A common example of environment data is input that is read from a
--- file.  Another is a large data structure constructed in-place.
---
--- By deferring the creation of an environment when its associated
--- benchmarks need the its, we avoid two problems that this strategy
--- caused:
---
--- * Memory pressure distorted the results of unrelated benchmarks.
---   If one benchmark needed e.g. a gigabyte-sized input, it would
---   force the garbage collector to do extra work when running some
---   other benchmark that had no use for that input.  Since the data
---   created by an environment is only available when it is in scope,
---   it should be garbage collected before other benchmarks are run.
---
--- * The time cost of generating all needed inputs could be
---   significant in cases where no inputs (or just a few) were really
---   needed.  This occurred often, for instance when just one out of a
---   large suite of benchmarks was run, or when a user would list the
---   collection of benchmarks without running any.
---
--- __Creation.__ An environment is created right before its related
--- benchmarks are run.  The 'IO' action that creates the environment
--- is run, then the newly created environment is evaluated to normal
--- form (hence the 'NFData' constraint) before being passed to the
--- function that receives the environment.
---
--- __Complex environments.__ If you need to create an environment that
--- contains multiple values, simply pack the values into a tuple.
---
--- __Lazy pattern matching.__ In situations where a \"real\"
--- environment is not needed, e.g. if a list of benchmark names is
--- being generated, @undefined@ will be passed to the function that
--- receives the environment.  This avoids the overhead of generating
--- an environment that will not actually be used.
---
--- The function that receives the environment must use lazy pattern
--- matching to deconstruct the tuple, as use of strict pattern
--- matching will cause a crash if @undefined@ is passed in.
---
--- __Example.__ This program runs benchmarks in an environment that
--- contains two values.  The first value is the contents of a text
--- file; the second is a string.  Pay attention to the use of a lazy
--- pattern to deconstruct the tuple in the function that returns the
--- benchmarks to be run.
---
--- > setupEnv = do
--- >   let small = replicate 1000 (1 :: Int)
--- >   big <- map length . words <$> readFile "/usr/dict/words"
--- >   return (small, big)
--- >
--- > main = defaultMain [
--- >    -- notice the lazy pattern match here!
--- >    env setupEnv $ \ ~(small,big) -> bgroup "main" [
--- >    bgroup "small" [
--- >      bench "length" $ whnf length small
--- >    , bench "length . filter" $ whnf (length . filter (==1)) small
--- >    ]
--- >  ,  bgroup "big" [
--- >      bench "length" $ whnf length big
--- >    , bench "length . filter" $ whnf (length . filter (==1)) big
--- >    ]
--- >  ] ]
---
--- __Discussion.__ The environment created in the example above is
--- intentionally /not/ ideal.  As Haskell's scoping rules suggest, the
--- variable @big@ is in scope for the benchmarks that use only
--- @small@.  It would be better to create a separate environment for
--- @big@, so that it will not be kept alive while the unrelated
--- benchmarks are being run.
-env :: NFData env =>
-       IO env
-    -- ^ Create the environment.  The environment will be evaluated to
-    -- normal form before being passed to the benchmark.
-    -> (env -> Benchmark)
-    -- ^ Take the newly created environment and make it available to
-    -- the given benchmarks.
-    -> Benchmark
-env alloc = Environment alloc noop
-
--- | Same as `env`, but but allows for an additional callback
--- to clean up the environment. Resource clean up is exception safe, that is,
--- it runs even if the 'Benchmark' throws an exception.
-envWithCleanup
-    :: NFData env
-    => IO env
-    -- ^ Create the environment.  The environment will be evaluated to
-    -- normal form before being passed to the benchmark.
-    -> (env -> IO a)
-    -- ^ Clean up the created environment.
-    -> (env -> Benchmark)
-    -- ^ Take the newly created environment and make it available to
-    -- the given benchmarks.
-    -> Benchmark
-envWithCleanup = Environment
-
--- | Create a Benchmarkable where a fresh environment is allocated for every
--- batch of runs of the benchmarkable.
---
--- The environment is evaluated to normal form before the benchmark is run.
---
--- When using 'whnf', 'whnfIO', etc. Gauge creates a 'Benchmarkable'
--- whichs runs a batch of @N@ repeat runs of that expressions. Gauge may
--- run any number of these batches to get accurate measurements. Environments
--- created by 'env' and 'envWithCleanup', are shared across all these batches
--- of runs.
---
--- This is fine for simple benchmarks on static input, but when benchmarking
--- IO operations where these operations can modify (and especially grow) the
--- environment this means that later batches might have their accuracy effected
--- due to longer, for example, longer garbage collection pauses.
---
--- An example: Suppose we want to benchmark writing to a Chan, if we allocate
--- the Chan using environment and our benchmark consists of @writeChan env ()@,
--- the contents and thus size of the Chan will grow with every repeat. If
--- Gauge runs a 1,000 batches of 1,000 repeats, the result is that the
--- channel will have 999,000 items in it by the time the last batch is run.
--- Since GHC GC has to copy the live set for every major GC this means our last
--- set of writes will suffer a lot of noise of the previous repeats.
---
--- By allocating a fresh environment for every batch of runs this function
--- should eliminate this effect.
-perBatchEnv
-    :: (NFData env, NFData b)
-    => (Int64 -> IO env)
-    -- ^ Create an environment for a batch of N runs. The environment will be
-    -- evaluated to normal form before running.
-    -> (env -> IO b)
-    -- ^ Function returning the IO action that should be benchmarked with the
-    -- newly generated environment.
-    -> Benchmarkable
-perBatchEnv alloc = perBatchEnvWithCleanup alloc (const noop)
-
--- | Same as `perBatchEnv`, but but allows for an additional callback
--- to clean up the environment. Resource clean up is exception safe, that is,
--- it runs even if the 'Benchmark' throws an exception.
-perBatchEnvWithCleanup
-    :: (NFData env, NFData b)
-    => (Int64 -> IO env)
-    -- ^ Create an environment for a batch of N runs. The environment will be
-    -- evaluated to normal form before running.
-    -> (Int64 -> env -> IO ())
-    -- ^ Clean up the created environment.
-    -> (env -> IO b)
-    -- ^ Function returning the IO action that should be benchmarked with the
-    -- newly generated environment.
-    -> Benchmarkable
-perBatchEnvWithCleanup alloc clean work
-    = Benchmarkable alloc clean (impure rnf . work) False
-
--- | Create a Benchmarkable where a fresh environment is allocated for every
--- run of the operation to benchmark. This is useful for benchmarking mutable
--- operations that need a fresh environment, such as sorting a mutable Vector.
---
--- As with 'env' and 'perBatchEnv' the environment is evaluated to normal form
--- before the benchmark is run.
---
--- This introduces extra noise and result in reduce accuracy compared to other
--- Gauge benchmarks. But allows easier benchmarking for mutable operations
--- than was previously possible.
-perRunEnv
-    :: (NFData env, NFData b)
-    => IO env
-    -- ^ Action that creates the environment for a single run.
-    -> (env -> IO b)
-    -- ^ Function returning the IO action that should be benchmarked with the
-    -- newly genereted environment.
-    -> Benchmarkable
-perRunEnv alloc = perRunEnvWithCleanup alloc noop
-
--- | Same as `perRunEnv`, but but allows for an additional callback
--- to clean up the environment. Resource clean up is exception safe, that is,
--- it runs even if the 'Benchmark' throws an exception.
-perRunEnvWithCleanup
-    :: (NFData env, NFData b)
-    => IO env
-    -- ^ Action that creates the environment for a single run.
-    -> (env -> IO ())
-    -- ^ Clean up the created environment.
-    -> (env -> IO b)
-    -- ^ Function returning the IO action that should be benchmarked with the
-    -- newly genereted environment.
-    -> Benchmarkable
-perRunEnvWithCleanup alloc clean work = bm { perRun = True }
-  where
-    bm = perBatchEnvWithCleanup (const alloc) (const clean) work
-
--- | Create a single benchmark.
-bench :: String                 -- ^ A name to identify the benchmark.
-      -> Benchmarkable          -- ^ An activity to be benchmarked.
-      -> Benchmark
-bench = Benchmark
-
--- | Group several benchmarks together under a common name.
-bgroup :: String                -- ^ A name to identify the group of benchmarks.
-       -> [Benchmark]           -- ^ Benchmarks to group under this name.
-       -> Benchmark
-bgroup = BenchGroup
-
--- | Add the given prefix to a name.  If the prefix is empty, the name
--- is returned unmodified.  Otherwise, the prefix and name are
--- separated by a @\'\/\'@ character.
-addPrefix :: String             -- ^ Prefix.
-          -> String             -- ^ Name.
-          -> String
-addPrefix ""  desc = desc
-addPrefix pfx desc = pfx ++ '/' : desc
-
--- | Retrieve the names of all benchmarks.  Grouped benchmarks are
--- prefixed with the name of the group they're in.
-benchNames :: Benchmark -> [String]
-benchNames (Environment _ _ b) = benchNames (b undefined)
-benchNames (Benchmark d _)   = [d]
-benchNames (BenchGroup d bs) = map (addPrefix d) . concatMap benchNames $ bs
-
-instance Show Benchmark where
-    show (Environment _ _ b) = "Environment _ _" ++ show (b undefined)
-    show (Benchmark d _)   = "Benchmark " ++ show d
-    show (BenchGroup d _)  = "BenchGroup " ++ show d
